// src/app/children/cabinet/children/dashboard/interfaces/vacancies-filters.interface.ts
export interface IVacanciesAppliedFilters {
    name: string;
    salary: string;
    city: string;
    employment: string;
    experience: string;
}

export interface IVacanciesFilters {
    name: string[];
    salary: string[];
    city: string[];
    employment: string[];
    experience: string[];
}